+++ packages/neuron-ui/src/components/AmendPendingTransactionDialog/index.tsx
@@ -3,12 +3,11 @@ import { useTranslation } from 'react-i18next'
 import { useState as useGlobalState } from 'states'
 import TextField from 'widgets/TextField'
 import Dialog from 'widgets/Dialog'
-import { MIN_AMOUNT } from 'utils/const'
-import { scriptToAddress } from '@nervosnetwork/ckb-sdk-utils'
+import { MIN_AMOUNT, DAO_DATA } from 'utils/const'
 import { isMainnet as isMainnetUtil, localNumberFormatter, shannonToCKBFormatter } from 'utils'
 import AlertDialog from 'widgets/AlertDialog'
 import styles from './amendPendingTransactionDialog.module.scss'
-import { useInitialize } from './hooks'
+import { useInitialize, useOutputs } from './hooks'
 
 const AmendPendingTransactionDialog = ({ tx, onClose }: { tx: State.Transaction; onClose: () => void }) => {
   const {
@@ -42,6 +41,14 @@ const AmendPendingTransactionDialog = ({ tx, onClose }: { tx: State.Transaction;
     onClose,
   })
 
+  const { items, lastOutputsCapacity } = useOutputs({
+    transaction,
+    isMainnet,
+    addresses,
+    sUDTAccounts,
+    fee,
+  })
+
   const priceError = useMemo(() => {
     return Number(price) < Number(minPrice) ? t('price-switch.errorTip', { minPrice }) : null
   }, [price, minPrice])
@@ -58,92 +65,13 @@ const AmendPendingTransactionDialog = ({ tx, onClose }: { tx: State.Transaction;
     [setPrice]
   )
 
-  const getLastOutputAddress = (outputs: State.DetailedOutput[]) => {
-    if (outputs.length === 1) {
-      return scriptToAddress(outputs[0].lock, isMainnet)
-    }
-
-    const change = outputs.find(output => {
-      const address = scriptToAddress(output.lock, isMainnet)
-      return !!addresses.find(item => item.address === address && item.type === 1)
-    })
-
-    if (change) {
-      return scriptToAddress(change.lock, isMainnet)
-    }
-
-    const receive = outputs.find(output => {
-      const address = scriptToAddress(output.lock, isMainnet)
-      return !!addresses.find(item => item.address === address && item.type === 0)
-    })
-    if (receive) {
-      return scriptToAddress(receive.lock, isMainnet)
-    }
-
-    const sudt = outputs.find(output => {
-      const address = scriptToAddress(output.lock, isMainnet)
-      return !!sUDTAccounts.find(item => item.address === address)
-    })
-    if (sudt) {
-      return scriptToAddress(sudt.lock, isMainnet)
-    }
-    return ''
-  }
-
-  const items: {
-    address: string
-    amount: string
-    capacity: string
-    isLastOutput: boolean
-    output: State.DetailedOutput
-  }[] = useMemo(() => {
-    if (transaction && transaction.outputs.length) {
-      const lastOutputAddress = getLastOutputAddress(transaction.outputs)
-      return transaction.outputs.map(output => {
-        const address = scriptToAddress(output.lock, isMainnet)
-        return {
-          capacity: output.capacity,
-          address,
-          output,
-          amount: shannonToCKBFormatter(output.capacity || '0'),
-          isLastOutput: address === lastOutputAddress,
-        }
-      })
-    }
-    return []
-  }, [transaction?.outputs])
-
-  const outputsCapacity = useMemo(() => {
-    const outputList = items.filter(item => !item.isLastOutput)
-    return outputList.reduce((total, cur) => {
-      if (Number.isNaN(+(cur.capacity || ''))) {
-        return total
-      }
-      return total + BigInt(cur.capacity || '0')
-    }, BigInt(0))
-  }, [items])
-
-  const lastOutputsCapacity = useMemo(() => {
-    if (transaction) {
-      const inputsCapacity = transaction.inputs.reduce((total, cur) => {
-        if (Number.isNaN(+(cur.capacity || ''))) {
-          return total
-        }
-        return total + BigInt(cur.capacity || '0')
-      }, BigInt(0))
-
-      return inputsCapacity - outputsCapacity - fee
-    }
-    return -1
-  }, [transaction, fee, outputsCapacity])
-
   useEffect(() => {
     if (transaction) {
       const outputs = items.map(item => {
         const capacity = item.isLastOutput ? lastOutputsCapacity.toString() : item.capacity
-        if (item.output.data === '0x0000000000000000') {
+        if (item.output.data === DAO_DATA) {
           // eslint-disable-next-line no-param-reassign
-          item.output.daoData = '0x0000000000000000'
+          item.output.daoData = DAO_DATA
         }
         return {
           ...item.output,
