+++ packages/neuron-ui/src/components/AmendSUDTSend/index.tsx
@@ -18,6 +18,7 @@ import {
 } from 'utils'
 import { DEFAULT_SUDT_FIELDS } from 'utils/const'
 import AlertDialog from 'widgets/AlertDialog'
+import { useOutputs } from 'components/AmendPendingTransactionDialog/hooks'
 import styles from './amendSUDTSend.module.scss'
 import { useInitialize } from './hooks'
 
@@ -61,6 +62,14 @@ const AmendSUDTSend = () => {
     t,
   })
 
+  const { items, lastOutputsCapacity } = useOutputs({
+    transaction,
+    isMainnet,
+    addresses,
+    sUDTAccounts,
+    fee,
+  })
+
   const priceError = useMemo(() => {
     return Number(price || '0') < Number(minPrice) ? t('price-switch.errorTip', { minPrice }) : null
   }, [price, minPrice])
@@ -94,81 +103,6 @@ const AmendSUDTSend = () => {
     }
     return scriptToAddress(transaction?.outputs[0].lock, isMainnet)
   }, [transaction?.outputs])
-
-  const getLastOutputAddress = (outputs: State.DetailedOutput[]) => {
-    const change = outputs.find(output => {
-      const address = scriptToAddress(output.lock, isMainnet)
-
-      return !!addresses.find(item => item.address === address && item.type === 1)
-    })
-    if (change) {
-      return scriptToAddress(change.lock, isMainnet)
-    }
-
-    const receive = outputs.find(output => {
-      const address = scriptToAddress(output.lock, isMainnet)
-      return !!addresses.find(item => item.address === address && item.type === 0)
-    })
-    if (receive) {
-      return scriptToAddress(receive.lock, isMainnet)
-    }
-
-    const sudt = outputs.find(output => {
-      const address = scriptToAddress(output.lock, isMainnet)
-      return !!sUDTAccounts.find(item => item.address === address)
-    })
-    if (sudt) {
-      return scriptToAddress(sudt.lock, isMainnet)
-    }
-    return ''
-  }
-
-  const items: {
-    address: string
-    amount: string
-    capacity: string
-    isLastOutput: boolean
-    output: State.DetailedOutput
-  }[] = useMemo(() => {
-    if (transaction && transaction.outputs.length) {
-      const lastOutputAddress = getLastOutputAddress(transaction.outputs)
-      return transaction.outputs.map(output => {
-        const address = scriptToAddress(output.lock, isMainnet)
-        return {
-          capacity: output.capacity,
-          address,
-          output,
-          amount: shannonToCKBFormatter(output.capacity || '0'),
-          isLastOutput: address === lastOutputAddress,
-        }
-      })
-    }
-    return []
-  }, [transaction?.outputs])
-
-  const outputsCapacity = useMemo(() => {
-    const outputList = items.filter(item => !item.isLastOutput)
-    return outputList.reduce((total, cur) => {
-      if (Number.isNaN(+(cur.capacity || ''))) {
-        return total
-      }
-      return total + BigInt(cur.capacity || '0')
-    }, BigInt(0))
-  }, [items])
-
-  const lastOutputsCapacity = useMemo(() => {
-    if (transaction) {
-      const inputsCapacity = transaction.inputs.reduce((total, cur) => {
-        if (Number.isNaN(+(cur.capacity || ''))) {
-          return total
-        }
-        return total + BigInt(cur.capacity || '0')
-      }, BigInt(0))
-
-      return inputsCapacity - outputsCapacity - fee
-    }
-    return -1
-  }, [transaction, fee, outputsCapacity])
 
   useEffect(() => {
     if (transaction) {
