+++ packages/neuron-ui/src/services/remote/offline.ts
@@ -45,7 +45,6 @@ export const exportTransactionAsJSON = remoteApi<OfflineSignJSON, void>('export-
 export const signTransactionOnly = remoteApi<OfflineSignJSON, void>('sign-transaction-only')
 export const broadcastTransaction = remoteApi<BroadcastProps, void>('broadcast-transaction-only')
 export const broadcastSignedTransaction = remoteApi<SignedTransaction, string>('broadcast-signed-transaction')
-export const getTransactionSize = remoteApi<State.DetailedTransaction, string>('get-transaction-size')
 export const signAndExportTransaction = remoteApi<SignProps, { filePath: string; json: OfflineSignJSON }>(
   'sign-and-export-transaction'
 )
