+++ packages/neuron-ui/src/services/remote/remoteApiWrapper.ts
@@ -139,7 +139,6 @@ type Action =
   | 'sign-transaction-only'
   | 'broadcast-transaction-only'
   | 'broadcast-signed-transaction'
-  | 'get-transaction-size'
   | 'sign-and-export-transaction'
   | 'sign-and-broadcast-transaction'
   // nft
