+++ packages/neuron-ui/src/utils/const.ts
@@ -83,3 +83,5 @@ export const ADDRESS_MIN_LENGTH = 86
 export const ADDRESS_HEAD_TAIL_LENGTH = 34
 
 export const PlaceHolderArgs = `0x${'00'.repeat(21)}`
+
+export const DAO_DATA = '0x0000000000000000'
